Move rainmachine coordinator to separate module
